+++ src/main/java/roomescape/admin/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.admin;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,13 +6,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import roomescape.reservation.Reservation;
+import roomescape.reservation.ReservationController;
 
 @Controller
 @RequestMapping("/admin")
 public class AdminController {
-
     private final ReservationController reservationController;
 
     @Autowired
