+++ src/main/java/roomescape/reservation/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -8,9 +8,6 @@ public class Reservation {
     private String name;
     private LocalDate date;
     private LocalTime time;
-
-    public Reservation() {
-    }
 
     public Reservation(Long id, String name, LocalDate date, LocalTime time) {
         this.id = id;
