+++ src/main/java/roomescape/reservation/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
